refactor(workflows): split node-config-drawer into field lib + component modules
Pure relocation (976 -> 561 lines): the field/coercion helpers, label
tables, and form styling constants move to lib/node-config-fields.ts
(zero JSX, unit-testable, canvas-editor-state precedent); DrawerHeader
and ParallelChildrenEditor become sibling components. JSX is
byte-identical where it moved; only imports and export keywords changed.

Also migrates the workflow-options 'let cancelled' effect (the FW6
straggler) to useJsonFetch from @makinbakin/sdk/hooks — a clean
single-JSON-GET fit; options derive via useMemo, error/empty behavior
unchanged. The drawer test's fetch assertion now allows useJsonFetch's
abort-signal options argument.
