+++ plugins/workflows/components/node-config-drawer.tsx
@@ -7,10 +7,13 @@
  * generated from the selected node type's `formFields` metadata and
  * validates the merged step payload against the node type's Zod schema
  * before Apply, so the drawer and the loader cannot drift.
+ *
+ * Pure field/coercion helpers live in `lib/node-config-fields`; the header
+ * bar and parallel-children editor are sibling components (FW4 split).
  */
 
-import { useEffect, useMemo, useState } from 'react'
-import { ArrowDown, ArrowUp, Plus, Trash2, X } from 'lucide-react'
+import { useMemo, useState } from 'react'
+import { Trash2 } from 'lucide-react'
 import { Button } from "@makinbakin/sdk/ui"
 import { Input } from "@makinbakin/sdk/ui"
 import { Textarea } from "@makinbakin/sdk/ui"
@@ -24,7 +27,31 @@ import {
   SelectValue,
 } from "@makinbakin/sdk/ui"
 import { AgentSelect } from "@makinbakin/sdk/components"
+import { useJsonFetch } from '@makinbakin/sdk/hooks'
 import { getNodeType, type FormField } from '@bakin/core/workflows/node-type-registry'
+
+import {
+  type ParallelChildRow,
+  type WorkflowSelectOption,
+  isDrawerEditableField,
+  fieldInitialValue,
+  coerceFieldValue,
+  normalizeParallelChildren,
+  isMissingRequiredField,
+  requiredFieldMessage,
+  schemaIssueMessage,
+  fieldLabel,
+  fieldHelpText,
+  fieldPlaceholder,
+  stepKindLabel,
+  FIELD_GROUP_CLASS,
+  FIELD_LABEL_CLASS,
+  FIELD_HELP_CLASS,
+  CONTROL_CLASS,
+  TEXTAREA_CLASS,
+} from '../lib/node-config-fields'
+import { DrawerHeader } from './node-config-drawer-header'
+import { ParallelChildrenEditor } from './parallel-children-editor'
 
 export interface NodeConfigDrawerProps {
   /** Step currently under edit, or null when nothing selected. */
@@ -44,216 +71,6 @@ export interface NodeConfigDrawerProps {
   existingStepIds?: string[]
   reservedStepIds?: string[]
 }
-
-type ParallelChildRow = {
-  id: string
-  type: string
-  label: string
-  [k: string]: unknown
-}
-
-type WorkflowSelectOption = {
-  id: string
-  name: string
-  disabled?: boolean
-}
-
-function isRecord(value: unknown): value is Record<string, unknown> {
-  return typeof value === 'object' && value !== null && !Array.isArray(value)
-}
-
-function isDrawerEditableField(kind: string, field: FormField): boolean {
-  return field.name !== 'dependsOn' && !(kind === 'parallel' && field.name === 'steps')
-}
-
-function coerceListInput(value: string): string[] {
-  return value
-    .split(',')
-    .map((s) => s.trim())
-    .filter(Boolean)
-}
-
-function fieldInitialValue(step: Record<string, unknown>, field: FormField): unknown {
-  const raw = step[field.name]
-  if (raw === undefined || raw === null) {
-    if (field.type === 'boolean') return false
-    if (field.type === 'list') return ''
-    return ''
-  }
-  if (field.name === 'on_reject' && isRecord(raw)) return raw.goto ?? ''
-  if (field.type === 'list' && Array.isArray(raw)) return raw.join(', ')
-  if (field.type === 'text' && isRecord(raw)) return JSON.stringify(raw, null, 2)
-  return raw
-}
-
-function coerceFieldValue(step: Record<string, unknown>, field: FormField, value: unknown): unknown {
-  if (field.name === 'on_reject') {
-    if (typeof value !== 'string' || value.trim().length === 0) return undefined
-    return {
-      ...(isRecord(step.on_reject) ? step.on_reject : {}),
-      goto: value.trim(),
-    }
-  }
-  if (field.name === 'content' && field.type === 'text' && typeof value === 'string') {
-    const trimmed = value.trim()
-    if (trimmed.length === 0) return undefined
-    let parsed: unknown
-    try {
-      parsed = JSON.parse(trimmed)
-    } catch {
-      throw new Error('Enter valid JSON object content.')
-    }
-    if (!isRecord(parsed)) {
-      throw new Error('Enter a JSON object such as {"summary": "..."}.')
-    }
-    return parsed
-  }
-  if (field.type === 'list' && typeof value === 'string') return coerceListInput(value)
-  return value
-}
-
-function normalizeParallelChildren(step: Record<string, unknown> | null): ParallelChildRow[] {
-  if (!step || !Array.isArray(step.steps)) return []
-  return step.steps
-    .filter(isRecord)
-    .map((child) => ({
-      ...child,
-      id: typeof child.id === 'string' ? child.id : '',
-      type: typeof child.type === 'string' ? child.type : 'agent',
-      label: typeof child.label === 'string' ? child.label : String(child.id ?? ''),
-    }))
-}
-
-function nextChildId(children: ParallelChildRow[]): string {
-  const existing = new Set(children.map((child) => child.id))
-  for (let i = 1; i < 1000; i++) {
-    const id = `child-${i}`
-    if (!existing.has(id)) return id
-  }
-  return `child-${Date.now()}`
-}
-
-function isMissingRequiredField(field: FormField, value: unknown): boolean {
-  if (!field.required) return false
-  if (value === undefined || value === null) return true
-  if (typeof value === 'string') return value.trim().length === 0
-  if (Array.isArray(value)) return value.length === 0
-  return false
-}
-
-function requiredFieldMessage(field: FormField): string {
-  if (field.type === 'agent') return 'Choose an agent.'
-  if (field.name === 'workflow_id') return 'Select a workflow.'
-  if (field.type === 'select') return `Select ${fieldLabel(field).toLowerCase()}.`
-  return `Enter ${fieldLabel(field).toLowerCase()}.`
-}
-
-function schemaIssueMessage(field: FormField, message: string): string {
-  if (message.toLowerCase().includes('required')) return requiredFieldMessage(field)
-  return message
-}
-
-function humanizeIdentifier(name: string): string {
-  return name
-    .replace(/([a-z0-9])([A-Z])/g, '$1 $2')
-    .replace(/[_-]+/g, ' ')
-    .replace(/\s+/g, ' ')
-    .trim()
-    .replace(/\b\w/g, (char) => char.toUpperCase())
-}
-
-const FIELD_LABELS: Record<string, string> = {
-  agent: 'Agent',
-  skill: 'Skill instructions',
-  task: 'Task brief',
-  description: 'Detailed instructions',
-  deny_tools: 'Denied tools',
-  approval_required: 'Require approval',
-  on_approve: 'Approved path',
-  on_reject: 'Rejected path',
-  preview: 'Preview fields',
-  channels: 'Notification channels',
-  content: 'Output content',
-  schedule: 'Schedule',
-  workflow_id: 'Nested workflow',
-  title: 'Task title',
-  column: 'Starting column',
-  workflowId: 'Attached workflow',
-  availableAt: 'Available after',
-  dueAt: 'Due date',
-}
-
-const FIELD_HELP_TEXT: Record<string, string> = {
-  agent: 'Choose who should run this step. Use Assigned agent to reuse the task assignee.',
-  skill: 'Optional skill or instruction bundle to load before the agent works.',
-  task: 'Short, concrete instruction for the agent. Use one or two sentences.',
-  description: 'Longer context, constraints, or acceptance criteria for this step.',
-  deny_tools: 'Optional comma-separated tool names the agent may not call during this step.',
-  approval_required: 'Require a human decision before the workflow can continue.',
-  on_approve: 'Step ID to run after approval. Use done when this approval completes the workflow.',
-  on_reject: 'Step ID to return to after rejection. Leave blank when rejection should not rewind.',
-  preview: 'Comma-separated output keys to show in the approval preview.',
-  channels: 'Comma-separated channels or destinations for the final output.',
-  content: 'Structured output content. Enter JSON when this step needs fixed output keys.',
-  schedule: 'Cron schedule for recurring output delivery.',
-  workflow_id: 'Select the workflow that this nested workflow step should run.',
-  title: 'Title for the task this workflow will create.',
-  column: 'Board column where the created task should start.',
-  workflowId: 'Optional workflow ID to attach to the created task.',
-  availableAt: 'Optional ISO timestamp before which the task should not dispatch.',
-  dueAt: 'Optional ISO timestamp for the desired completion time.',
-}
-
-const FIELD_PLACEHOLDERS: Record<string, string> = {
-  skill: 'brand-voice',
-  task: 'Write a concise post caption for the approved brief.',
-  description: 'Include caption text, target platform notes, hashtags, and any required mentions.',
-  deny_tools: 'web.run, shell.exec',
-  on_approve: 'review-copy or done',
-  on_reject: 'revise-copy',
-  preview: 'caption, hashtags, mentions',
-  channels: 'general, announcements',
-  content: '{"summary": "...", "assetPath": "..."}',
-  schedule: '0 9 * * 1',
-  title: 'Draft launch announcement',
-  workflowId: 'social-post',
-  availableAt: '2026-05-25T14:00:00Z',
-  dueAt: '2026-05-26T18:00:00Z',
-}
-
-const STEP_KIND_LABELS: Record<string, string> = {
-  agent: 'Agent step',
-  gate: 'Approval gate',
-  output: 'Completion step',
-  workflow: 'Nested workflow',
-  parallel: 'Parallel group',
-  createTask: 'Task creation step',
-}
-
-function fieldLabel(field: FormField): string {
-  return FIELD_LABELS[field.name] ?? humanizeIdentifier(field.name)
-}
-
-function fieldHelpText(field: FormField): string | undefined {
-  return FIELD_HELP_TEXT[field.name] ?? field.description
-}
-
-function fieldPlaceholder(field: FormField): string | undefined {
-  return FIELD_PLACEHOLDERS[field.name]
-}
-
-function stepKindLabel(kind: string): string {
-  if (!kind) return 'Unknown step'
-  if (STEP_KIND_LABELS[kind]) return STEP_KIND_LABELS[kind]
-  const localKind = kind.includes('.') ? kind.split('.').slice(1).join(' ') : kind
-  return `${humanizeIdentifier(localKind)} step`
-}
-
-const FIELD_GROUP_CLASS = 'space-y-2.5'
-const FIELD_LABEL_CLASS = 'text-sm font-medium leading-none'
-const FIELD_HELP_CLASS = 'text-xs leading-relaxed text-muted-foreground/60'
-const CONTROL_CLASS = 'min-h-10 text-sm'
-const TEXTAREA_CLASS = 'min-h-24 text-sm leading-relaxed'
 
 export function NodeConfigDrawer({
   step,
@@ -287,7 +104,6 @@ export function NodeConfigDrawer({
     return seeded
   })
   const [parallelChildren, setParallelChildren] = useState<ParallelChildRow[]>(() => normalizeParallelChildren(step))
-  const [workflowOptions, setWorkflowOptions] = useState<WorkflowSelectOption[]>([])
   const [fieldErrors, setFieldErrors] = useState<Record<string, string>>({})
   const [idError, setIdError] = useState<string | null>(null)
   const [labelError, setLabelError] = useState<string | null>(null)
@@ -305,35 +121,18 @@ export function NodeConfigDrawer({
     onDirtyChange?.(true)
   }
 
-  useEffect(() => {
-    if (!needsWorkflowOptions) return
-    let cancelled = false
-
-    async function loadWorkflowOptions() {
-      try {
-        const res = await fetch('/api/plugins/workflows/definitions?includeDisabled=1')
-        if (!res.ok) return
-        const data = (await res.json()) as {
-          templates?: Array<{ filename: string; name: string; disabled?: boolean }>
-        }
-        if (cancelled) return
-        setWorkflowOptions(
-          (data.templates ?? []).map((template) => ({
-            id: template.filename,
-            name: template.name,
-            disabled: template.disabled,
-          })),
-        )
-      } catch {
-        if (!cancelled) setWorkflowOptions([])
-      }
-    }
-
-    void loadWorkflowOptions()
-    return () => {
-      cancelled = true
-    }
-  }, [needsWorkflowOptions])
+  const { data: workflowOptionsData } = useJsonFetch<{
+    templates?: Array<{ filename: string; name: string; disabled?: boolean }>
+  }>(needsWorkflowOptions ? '/api/plugins/workflows/definitions?includeDisabled=1' : null)
+  const workflowOptions = useMemo<WorkflowSelectOption[]>(
+    () =>
+      (workflowOptionsData?.templates ?? []).map((template) => ({
+        id: template.filename,
+        name: template.name,
+        disabled: template.disabled,
+      })),
+    [workflowOptionsData],
+  )
 
   if (!step) return null
 
@@ -605,220 +404,6 @@ export function NodeConfigDrawer({
         </div>
       </div>
     </aside>
-  )
-}
-
-function DrawerHeader({
-  title,
-  subtitle,
-  onClose,
-}: {
-  title: string
-  subtitle?: string
-  onClose: () => void
-}) {
-  return (
-    <div className="flex items-center justify-between border-b border-border px-5 py-3">
-      <div className="flex flex-col leading-tight">
-        <span className="text-base font-medium">{title}</span>
-        {subtitle && <span className="text-xs text-muted-foreground/60">{subtitle}</span>}
-      </div>
-      <button
-        type="button"
-        aria-label="Close drawer"
-        className="rounded p-1 text-muted-foreground hover:bg-muted"
-        onClick={onClose}
-      >
-        <X className="size-4" />
-      </button>
-    </div>
-  )
-}
-
-function ParallelChildrenEditor({
-  childrenRows,
-  onChange,
-}: {
-  childrenRows: ParallelChildRow[]
-  onChange: (next: ParallelChildRow[]) => void
-}) {
-  const updateChild = (index: number, patch: Record<string, unknown>) => {
-    onChange(childrenRows.map((child, i) => (i === index ? { ...child, ...patch } : child)))
-  }
-  const removeChild = (index: number) => {
-    onChange(childrenRows.filter((_, i) => i !== index))
-  }
-  const moveChild = (index: number, delta: -1 | 1) => {
-    const nextIndex = index + delta
-    if (nextIndex < 0 || nextIndex >= childrenRows.length) return
-    const next = [...childrenRows]
-    const [moved] = next.splice(index, 1)
-    next.splice(nextIndex, 0, moved)
-    onChange(next)
-  }
-  const addChild = () => {
-    const id = nextChildId(childrenRows)
-    onChange([
-      ...childrenRows,
-      {
-        id,
-        type: 'agent',
-        label: id,
-        agent: '$assigned',
-      },
-    ])
-  }
-
-  return (
-    <div className="border-t border-border pt-5">
-      <div className="mb-4 flex items-center justify-between gap-2">
-        <span className="text-sm font-medium">Parallel child steps</span>
-        <Button
-          type="button"
-          variant="ghost"
-          size="sm"
-          onClick={addChild}
-          aria-label="Add child agent"
-        >
-          <Plus className="mr-1 size-3.5" /> Agent
-        </Button>
-      </div>
-      <div className="space-y-5">
-        {childrenRows.map((child, index) => {
-          const childId = child.id || `child-${index + 1}`
-          if (child.type !== 'agent') {
-            return (
-              <div key={`${childId}-${index}`} className="rounded border border-amber-500/40 bg-amber-500/10 p-3 text-xs leading-relaxed text-amber-200">
-                Child {childId} has unsupported type {child.type}; it is preserved read-only.
-              </div>
-            )
-          }
-          return (
-            <div key={`${childId}-${index}`} className="rounded border border-border p-4">
-              <div className="mb-4 flex items-center justify-between gap-2">
-                <span className="min-w-0 truncate text-sm font-medium">{child.label || childId}</span>
-                <div className="flex items-center gap-1">
-                  <Button
-                    type="button"
-                    variant="ghost"
-                    size="sm"
-                    aria-label={`Move child ${childId} up`}
-                    onClick={() => moveChild(index, -1)}
-                    disabled={index === 0}
-                  >
-                    <ArrowUp className="size-3.5" />
-                  </Button>
-                  <Button
-                    type="button"
-                    variant="ghost"
-                    size="sm"
-                    aria-label={`Move child ${childId} down`}
-                    onClick={() => moveChild(index, 1)}
-                    disabled={index === childrenRows.length - 1}
-                  >
-                    <ArrowDown className="size-3.5" />
-                  </Button>
-                  <Button
-                    type="button"
-                    variant="ghost"
-                    size="sm"
-                    aria-label={`Remove child ${childId}`}
-                    onClick={() => removeChild(index)}
-                  >
-                    <Trash2 className="size-3.5" />
-                  </Button>
-                </div>
-              </div>
-              <div className="grid gap-5">
-                <div className={FIELD_GROUP_CLASS}>
-                  <Label className={FIELD_LABEL_CLASS} htmlFor={`parallel-child-${index}-id`}>
-                    Child step ID
-                  </Label>
-                  <Input
-                    id={`parallel-child-${index}-id`}
-                    aria-label={`Child step ID for ${childId}`}
-                    value={child.id}
-                    placeholder="write-caption"
-                    className={CONTROL_CLASS}
-                    onChange={(e) => updateChild(index, { id: e.target.value })}
-                  />
-                  <p className={FIELD_HELP_CLASS}>
-                    Stable identifier used by workflow links.
-                  </p>
-                </div>
-                <div className={FIELD_GROUP_CLASS}>
-                  <Label className={FIELD_LABEL_CLASS} htmlFor={`parallel-child-${index}-label`}>
-                    Display name
-                  </Label>
-                  <Input
-                    id={`parallel-child-${index}-label`}
-                    aria-label={`Display name for child ${childId}`}
-                    value={child.label}
-                    placeholder="Write Caption"
-                    className={CONTROL_CLASS}
-                    onChange={(e) => updateChild(index, { label: e.target.value })}
-                  />
-                  <p className={FIELD_HELP_CLASS}>
-                    Human-readable name shown inside the parallel group.
-                  </p>
-                </div>
-                <div className={FIELD_GROUP_CLASS}>
-                  <Label className={FIELD_LABEL_CLASS}>Agent</Label>
-                  <AgentSelect
-                    value={typeof child.agent === 'string' ? child.agent : ''}
-                    onValueChange={(v) => updateChild(index, { agent: v || '' })}
-                    includeAssigned
-                    allowNone={false}
-                    className={CONTROL_CLASS}
-                  />
-                  <p className={FIELD_HELP_CLASS}>
-                    Choose who should run this child step.
-                  </p>
-                </div>
-                <div className={FIELD_GROUP_CLASS}>
-                  <Label className={FIELD_LABEL_CLASS} htmlFor={`parallel-child-${index}-skill`}>
-                    Skill instructions
-                  </Label>
-                  <Input
-                    id={`parallel-child-${index}-skill`}
-                    aria-label={`Skill instructions for child ${childId}`}
-                    value={typeof child.skill === 'string' ? child.skill : ''}
-                    placeholder="brand-voice"
-                    className={CONTROL_CLASS}
-                    onChange={(e) => updateChild(index, { skill: e.target.value || undefined })}
-                  />
-                  <p className={FIELD_HELP_CLASS}>
-                    Optional skill or instruction bundle to load first.
-                  </p>
-                </div>
-                <div className={FIELD_GROUP_CLASS}>
-                  <Label className={FIELD_LABEL_CLASS} htmlFor={`parallel-child-${index}-task`}>
-                    Task brief
-                  </Label>
-                  <Textarea
-                    id={`parallel-child-${index}-task`}
-                    aria-label={`Task brief for child ${childId}`}
-                    rows={3}
-                    value={typeof child.task === 'string' ? child.task : ''}
-                    placeholder="Draft the caption for this audience segment."
-                    className={TEXTAREA_CLASS}
-                    onChange={(e) => updateChild(index, { task: e.target.value || undefined })}
-                  />
-                  <p className={FIELD_HELP_CLASS}>
-                    Short, concrete instruction for this child agent.
-                  </p>
-                </div>
-              </div>
-            </div>
-          )
-        })}
-        {childrenRows.length === 0 && (
-          <div className="rounded border border-border p-3 text-xs leading-relaxed text-muted-foreground">
-            Add at least one child agent before saving this parallel step.
-          </div>
-        )}
-      </div>
-    </div>
   )
 }
